+++ rocky/reports/views/aggregate_report.py
@@ -7,6 +7,7 @@ from django.shortcuts import redirect
 from django.urls import reverse
 from django.utils.http import urlencode
 from django.utils.translation import gettext_lazy as _
+from tools.view_helpers import PostRedirect
 
 from reports.report_types.aggregate_organisation_report.report import AggregateOrganisationReport
 from reports.report_types.definitions import AggregateReport, MultiReport, Report
@@ -14,7 +15,6 @@ from reports.report_types.helpers import get_ooi_types_from_aggregate_report, ge
 from reports.views.base import (
     REPORTS_PRE_SELECTION,
     OOISelectionView,
-    PostRedirect,
     ReportBreadcrumbs,
     ReportPluginView,
     ReportTypeSelectionView,
