+++ rocky/tools/view_helpers.py
@@ -3,8 +3,9 @@ from datetime import date, datetime, timezone
 from typing import TypedDict
 from urllib.parse import urlencode, urlparse, urlunparse
 
-from django.http import HttpRequest
+from django.http import HttpRequest, HttpResponse
 from django.urls.base import reverse, reverse_lazy
+from django.utils.encoding import iri_to_uri
 from django.utils.translation import gettext_lazy as _
 
 from octopoes.models.types import OOI_TYPES
@@ -163,3 +164,11 @@ class ObjectsBreadcrumbsMixin(BreadcrumbsMixin):
                 "text": _("Objects"),
             }
         ]
+
+
+class PostRedirect(HttpResponse):
+    status_code = 307
+
+    def __init__(self, redirect_to):
+        super().__init__(self)
+        self["Location"] = iri_to_uri(redirect_to)
